+++ CRM/Searchactiondesigner/Form/Task/Contribution.php
@@ -29,7 +29,11 @@ class CRM_Searchactiondesigner_Form_Task_Contribution extends CRM_Contribute_For
 
   public function buildQuickForm() {
     CRM_Searchactiondesigner_Form_Task_Helper::buildQuickForm($this, $this->searchTaskId);
-    $this->addDefaultButtons(E::ts('Next'), 'upload');
+    $label = E::ts('Next');
+    if (!empty($this->searchTask['proceed_label'])) {
+      $label = $this->searchTask['proceed_label'];
+    }
+    $this->addDefaultButtons($label, 'upload');
   }
 
   public function postProcess() {